+++ src/investment/BaasInvestCode.ts
@@ -1,0 +1,26 @@
+export enum BaasInvestCode {
+  /**
+   * Created
+   */
+  BI0001 = "BI0001",
+  /**
+   * Confirmed
+   */
+  BI0002 = "BI0002",
+  /**
+   * Not confirmed
+   */
+  BI0003 = "BI0003",
+  /**
+   * Processed
+   */
+  BI0004 = "BI0004",
+  /**
+   * Done
+   */
+  BI0005 = "BI0005",
+  /**
+   * Insufficient balance
+   */
+  BI0006 = "BI0006",
+}