+++ src/investment/index.ts
@@ -1,0 +1,5 @@
+import {BaasInvestCode} from './BaasInvestCode';
+
+export default Object.freeze({
+  ...BaasInvestCode
+});